Completed manager node code for waiting containers

+++ middle-layer/manager_node/startup.sh
@@ -1,44 +1,80 @@
 #!/bin/bash
 
 
-# Gets the username
-export USER=$(curl -s http://$MANAGER_NODE:5000/api/instance/whoami/$UUID_f10)
+# If the container was left as waiting:
+if [ -f "/etc/wait.key" ]; then
+    
+    wait_key_file="/etc/wait.key"
+
+    # Gets the key
+    wait_key=$(head -n 1 "$wait_key_file")
+
+    # Gets the username
+    export USER=$(head -2 "$wait_key_file" | tail -1)
+
+    # Asks the user to enter the key value
+    printf "\nEnter the WAIT key (this process is done automatically): "
+    read user_provided_wait_key
+
+    if [ "$user_provided_wait_key" = "$wait_key" ]; then
+
+        # Deletes the wait file
+        curl -s http://0.0.0.0:3100/delete_wait
+        export GS=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/location)
+        GK2=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/new/single_use_token/$UUID_f10)
+        printf "\n\033[0;32mWelcome back\033[0m\n\n"
+
+    else
+        export USER="Empty"
+        printf "Access not allowed\n"
+        exit
+    fi
 
 
+    # No need to setup the rest of the container
 
-# If the user is not what is expected, it exits
-if [ "$USER" = "Empty" ]; then
-   printf "Access not allowed\n"
-   exit
+# New user
+else
+
+    # Gets the username
+    export USER=$(curl -s http://$MANAGER_NODE:5000/api/instance/whoami/$UUID_f10)
+
+    # If the user is not what is expected, it exits
+    if [ "$USER" = "Empty" ]; then
+       printf "Access not allowed\n"
+       exit
+    fi
+
+    # Gets the greyfish server location
+    export GS=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/location)
+
+    # Gets two greyfish keys
+    GK1=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/new/single_use_token/$UUID_f10)
+    GK2=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/new/single_use_token/$UUID_f10)
+
+
+    # Gets all the user files
+    curl -s http://$GS:2001/grey/get_all/$GK1/$USER > summary.tar.gz
+
+    # Checks that the previous data is actually tarred
+    if { tar ztf "summary.tar.gz" || tar tf "summary.tar.gz"; } >/dev/null 2>&1; then
+       tar -xzf summary.tar.gz
+       rm -f /home/gib/home/gib/.bash_logout
+       mv /home/gib/home/gib/* /home/gib
+       rm -rf /home/gib/home
+    fi
+
+    rm -f summary.tar.gz
+
+
 fi
-
-# Gets the greyfish server location
-export GS=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/location)
-
-# Gets two greyfish keys
-GK1=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/new/single_use_token/$UUID_f10)
-GK2=$(curl -s http://$MANAGER_NODE:5000/api/greyfish/new/single_use_token/$UUID_f10)
-
-
-# Gets all the user files
-curl -s http://$GS:2001/grey/get_all/$GK1/$USER > summary.tar.gz
-
-# Checks that the previous data is actually tarred
-if { tar ztf "summary.tar.gz" || tar tf "summary.tar.gz"; } >/dev/null 2>&1; then
-   tar -xzf summary.tar.gz
-   rm -f /home/gib/home/gib/.bash_logout
-   mv /home/gib/home/gib/* /home/gib
-   rm -rf /home/gib/home
-fi
-
-rm -f summary.tar.gz
 
 # Gets the project
 export PROJECT=$(curl -s http://$MANAGER_NODE:5000/api/project/name)
 
 # Changes the terminal prompt to [/PROJECT/USERNAME: ~] $ 
-
 PS1="$USER@$PROJECT: $PWD \$ " 
+
 
 
 function slurm_submit     {
