+++ new-wetty/bash_logout.sh
@@ -6,14 +6,19 @@
 if [ "$USER" != "root" ]; then
 
     if [ "$USER" != "Empty" ]; then
-        # Free instance
-        curl -s http://$MANAGER_NODE:5000/api/instance/freeme/$UUID_f10
 
-        # Creates a temporary directory and moves all the stuff in the current one there
-        tar -zcf summary.tar.gz  /home/gib
-        curl -F file=@summary.tar.gz  http://$GS:2002/grey/push_all/$GK2/$USER
+    	# If the user has left a waiting server
+    	if [ ! -f "/etc/wait.key" ]; then
 
-        # Deletes all user files that can be deleted
-        rm -rf /home/gib/*
+            # Free instance
+            curl -s http://$MANAGER_NODE:5000/api/instance/freeme/$UUID_f10
+
+            # Creates a temporary directory and moves all the stuff in the current one there
+            tar -zcf summary.tar.gz  /home/gib
+            curl -F file=@summary.tar.gz  http://$GS:2002/grey/push_all/$GK2/$USER
+
+            # Deletes all user files that can be deleted
+            rm -rf /home/gib/*
+        fi
     fi
 fi
